Calculate delivery date async
* Closes #958
