+++ app/assets/javascripts/deliveryDueAt.js
@@ -1,13 +1,23 @@
 $(document).ready(function() {
   calculateNewDate = function() {
-    $('.estimated-delivery-date').text("Save your changes to see updated delivery date and time");
+    auctionEndedAt = $("#auction_ended_at").val();
+    dueInDays = $("#auction_due_in_days").val();
+    hour = $("#auction_ended_at_1i").val();
+    minute = $("#auction_ended_at_2i").val();
+    amPm = $("#auction_ended_at_3i").val();
+    $.getJSON(
+      '/api/v0/business_day',
+      { day_count: dueInDays, date: auctionEndedAt, time: (hour + ":" + minute + " " + amPm)  },
+      function(data) {
+        $('.delivery-date').text(data.date);
+    });
   };
 
-  $( ".auction_due_in_days").change(function() {
+  $(".auction_due_in_days").change(function() {
     calculateNewDate();
   });
 
-  $( ".auction_ended_at").change(function() {
+  $(".ended_at").change(function() {
     calculateNewDate();
   });
 });
