+++ app/views/admin/auctions/_form.html.erb
@@ -5,7 +5,7 @@
     locals: { field: 'started', auction: auction } %>
 </fieldset>
 
-<fieldset class="datetime">
+<fieldset class="datetime ended_at">
   <legend>Auction ends</legend>
   <%= render partial: 'datetime_fields',
     locals: { field: 'ended', auction: auction } %>
