+++ features/step_definitions/admin_auction_form_steps.rb
@@ -71,7 +71,7 @@ When(/^I edit the new auction form$/) do
 end
 
 When(/^I change the auction end date$/) do
-  @end_day = DcTimePresenter.convert(3.days.from_now)
+  @end_day = 3.days.from_now
   fill_in "auction_ended_at", with: @end_day.strftime('%Y-%m-%d')
 end
 
@@ -79,6 +79,16 @@ Then(/^I should see an estimated delivery deadline of 12 business days from now$
   within('.estimated-delivery-date') do
     default_delivery_date = DefaultDateTime.new(12.business_days.from_now).convert
     expect(page).to have_content("Estimated delivery date #{DcTimePresenter.convert_and_format(default_delivery_date)}")
+  end
+end
+
+Then(/^I should see the updated delivery deadline$/) do
+  within('.estimated-delivery-date') do
+    delivery_date = DefaultDateTime.new(5.business_days.after(3.days.from_now)).convert
+    formatted_delivery_date = DcTimePresenter.convert_and_format(delivery_date)
+    expect(page).to have_content(
+      "Estimated delivery date #{formatted_delivery_date}"
+    )
   end
 end
