+++ spec/requests/api/v0/bids_spec.rb
